refactor(role): wordMuteLimit lookup を closure memoize + draft count を int64 比較に (#1029 PR-1 follow-up)
PR #1039 (#1029 PR-1) の review feedback 反映。

- `internal/api/i/handler.go`: `wordMuteLimit` の policy lookup を closure に
  集約。mutedWords / hardMutedWords 両方指定された request で同 policy を 2
  度 lookup していたのを 1 度に集約 (`canUpdateBioMedia` と同 pattern、
  upstream `policies ??= await ...` と等価)。
- `internal/api/notes/handler_drafts.go`: `int(count) >= limit` を
  `count >= int64(limit)` に書き換え、32-bit platform での overflow を防ぐ
  defensive cast に変更。production は 64-bit + limit は数百以下で発火しない
  が、`CountByUser` は `int64` を返すので比較側を合わせる方が正しい。

該当 2 パッケージ test green、`gofmt -s` / `go vet` clean。

+++ internal/api/notes/handler_drafts.go
@@ -59,7 +59,7 @@ func (h *Handler) DraftsCreate(c echo.Context) error {
 			if err != nil {
 				return apierr.JSONInternalError(c)
 			}
-			if int(count) >= limit {
+			if count >= int64(limit) {
 				return apierr.JSONTooManyNoteDrafts(c)
 			}
 		}
